requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.5.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.4
cloudinary==1.38.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.2
django-cors-headers==4.1.0
django-database-url==1.0.3
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==23.1
pandas==2.2.3
Pillow==10.0.0
psycopg2-binary
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.18.0
PyJWT==2.8.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
rave-python==1.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
